(fix) Renaming to correctly reflect the CONNECT-IP protocol name
