+++ src/lib.rs
@@ -2,6 +2,6 @@ pub mod client;
 pub mod server;
 pub mod socks;
 pub mod http;
-pub mod ip_connect;
+pub mod connect_ip;
 pub mod common;
